+++ packages/builders/src/index.ts
@@ -11,8 +11,27 @@ export * from './components/textInput/TextInput.js';
 export * as TextInputAssertions from './components/textInput/Assertions.js';
 export * from './interactions/modals/Modal.js';
 export * as ModalAssertions from './interactions/modals/Assertions.js';
-export * from './components/selectMenu/SelectMenu.js';
-export * from './components/selectMenu/SelectMenuOption.js';
+
+export * from './components/selectMenu/BaseSelectMenu.js';
+export * from './components/selectMenu/ChannelSelectMenu.js';
+export * from './components/selectMenu/MentionableSelectMenu.js';
+export * from './components/selectMenu/RoleSelectMenu.js';
+export * from './components/selectMenu/StringSelectMenu.js';
+// TODO: Remove those aliases in v2
+export {
+	/**
+	 * @deprecated Will be removed in the next major version, use {@link StringSelectMenuBuilder} instead.
+	 */
+	StringSelectMenuBuilder as SelectMenuBuilder,
+} from './components/selectMenu/StringSelectMenu.js';
+export {
+	/**
+	 * @deprecated Will be removed in the next major version, use {@link StringSelectMenuOptionBuilder} instead.
+	 */
+	StringSelectMenuOptionBuilder as SelectMenuOptionBuilder,
+} from './components/selectMenu/StringSelectMenuOption.js';
+export * from './components/selectMenu/StringSelectMenuOption.js';
+export * from './components/selectMenu/UserSelectMenu.js';
 
 export * as SlashCommandAssertions from './interactions/slashCommands/Assertions.js';
 export * from './interactions/slashCommands/SlashCommandBuilder.js';
@@ -40,3 +59,11 @@ export * from './util/componentUtil.js';
 export * from './util/normalizeArray.js';
 export * from './util/validation.js';
 export * from '@discordjs/util';
+
+/**
+ * The {@link https://github.com/discordjs/discord.js/blob/main/packages/builders/#readme | @discordjs/builders} version
+ * that you are currently using.
+ */
+// This needs to explicitly be `string` so it is not typed as a "const string" that gets injected by esbuild
+// eslint-disable-next-line @typescript-eslint/no-inferrable-types
+export const version: string = '[VI]{{inject}}[/VI]';