+++ packages/discord.js/src/structures/ContextMenuCommandInteraction.js
@@ -25,7 +25,7 @@ class ContextMenuCommandInteraction extends CommandInteraction {
     );
 
     /**
-     * The id of the target of the interaction
+     * The id of the target of this interaction
      * @type {Snowflake}
      */
     this.targetId = data.data.target_id;