+++ packages/scripts/src/index.ts
@@ -1,1 +1,2 @@
 export * from './generateIndex.js';
+export * from './createPackage.js';