+++ packages/ui/src/lib/components/discord/MessageEmbedTitle.tsx
@@ -1,0 +1,7 @@
+export interface IDiscordMessageEmbedTitle {
+	title: string;
+}
+
+export function DiscordMessageEmbedTitle({ title }: IDiscordMessageEmbedTitle) {
+	return <div className="mt-2 font-medium">{title}</div>;
+}