+++ packages/voice/src/index.ts
@@ -19,3 +19,11 @@ export {
 } from './VoiceConnection';
 
 export { type JoinConfig, getVoiceConnection, getVoiceConnections, getGroups } from './DataStore';
+
+/**
+ * The {@link https://github.com/discordjs/discord.js/blob/main/packages/voice/#readme | @discordjs/voice} version
+ * that you are currently using.
+ */
+// This needs to explicitly be `string` so it is not typed as a "const string" that gets injected by esbuild
+// eslint-disable-next-line @typescript-eslint/no-inferrable-types
+export const version: string = '[VI]{{inject}}[/VI]';