+++ packages/voice/src/util/generateDependencyReport.ts
@@ -33,10 +33,11 @@ function findPackageJSON(
  */
 function version(name: string): string {
 	try {
-		const pkg =
-			name === '@discordjs/voice'
-				? require('../../package.json')
-				: findPackageJSON(dirname(require.resolve(name)), name, 3);
+		if (name === '@discordjs/voice') {
+			return '[VI]{{inject}}[/VI]';
+		}
+
+		const pkg = findPackageJSON(dirname(require.resolve(name)), name, 3);
 		return pkg?.version ?? 'not found';
 	} catch {
 		return 'not found';